+++ graphql/queries/query.js
@@ -1,0 +1,20 @@
+// Prepare our GraphQL query
+const GET_ALL_POSTS = `
+  query {
+    postCollection {
+      edges {
+        node {
+          id
+          ownerId
+        }
+      }
+    }
+  }
+`
+
+export { GET_ALL_POSTS };
+
+
+
+
+
