+++ lib/supabaseClient.js
@@ -1,12 +1,28 @@
-import { createClient } from '@supabase/supabase-js';
+import { createClient as supabaseClient } from '@supabase/supabase-js';
 import { v4 as uuidv4 } from 'uuid';
-import { queryPost } from './graphql/queries/query';
+import { createClient as urqlClient, useQuery as urqlQuery } from 'urql';
+import { GET_ALL_POSTS } from '../graphql/queries/query';
 
+// Prepare API key and Authorization header
+const headers = {
+    apikey: process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
+    authorization: `Bearer: ${process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY}`
+}
+
+// Create GraphQL client
+// See: https://formidable.com/open-source/urql/docs/basics/react-preact/#setting-up-the-client
+
+export const supabaseGraphQLClient = urqlClient({
+    url: `${process.env.NEXT_PUBLIC_SUPABASE_URL}/grapqhl/v1`,
+    fetchOptions: function createFetchOptions() {
+        return { headers }
+    },
+})
 
 const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL
 const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
 
-export const supabase = createClient(supabaseUrl, supabaseAnonKey)
+export const supabase = supabaseClient(supabaseUrl, supabaseAnonKey)
 
 // Upload new post to Supabase
 export async function postToSupa(postToSave, toast) {
@@ -65,29 +81,6 @@ export async function postToSupa(postToSave, toast) {
 
         ];
 
-
-    console.log({
-        id: id,
-        owner: owner,
-        token_id: token_id,
-        owner_id: owner_id,
-        post_id: post_id,
-        description: description,
-        title: title,
-        media: media,
-        media_hash: media_hash,
-        media_type: media_type,
-        created_at: created_at,
-        updated_at: updated_at,
-        totalcharged: totalcharged,
-        extra: extra,
-        status: status,
-        comments: comments,
-    }
-
-
-    );
-
     // Insert into account
     const { data: aaccountData, error: accountError } = await supabase.from("account").insert([{
         id,
@@ -96,8 +89,6 @@ export async function postToSupa(postToSave, toast) {
         created_at,
         updated_at
     }]);
-
-    console.log({ aaccountData: aaccountData })
 
     //Insert into table `post`
     const { data: postData, error: postError } = await supabase.from("post").insert([{
@@ -108,8 +99,6 @@ export async function postToSupa(postToSave, toast) {
         created_at,
         updated_at
     }]);
-
-    console.log({ postData: postData })
 
     const { data: postNFT, error: postNFTError } = await supabase.from("post_nft").insert([{
         post_id,
@@ -129,7 +118,6 @@ export async function postToSupa(postToSave, toast) {
 
     }]);
 
-    console.log({ postNFT: postNFT })
     if (accountError) {
         toast(
             "error",
@@ -156,9 +144,16 @@ export async function postToSupa(postToSave, toast) {
     } else {
         
         console.log(" Uploaded successfully to Supabase", "success")
+        
+
+
+
+        console.log({GET_ALL_POSTS: GET_AL_POSTS})
         // redirect back to feed
     }
 }
+
+
 
 export async function fileUpload(file) {
     if (file) {
