+++ package.json
@@ -41,7 +41,6 @@
         "react": "17.0.2",
         "react-dom": "17.0.2",
         "react-icons": "^4.3.1",
-        "react-is": "^18.0.0",
         "sha.js": "^2.4.11",
         "timeago-react": "^3.0.4",
         "urql": "^2.2.0",
